# spec/openapi.yaml
openapi: 3.0.3
info:
  title: Book Lending API
  version: 1.0.0
  description: API specification for the book lending application.
paths:
  /api/books:
    get:
      operationId: GetAllBooks::BookService
      x-route-group: bookGroup::/api/books
      tags:
        - Books
      summary: Retrieve a list of all books
      responses:
        "200":
          description: A list of books
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Book"
    post:
      operationId: CreateBook::BookService
      x-route-group: bookGroup::/api/books
      tags:
        - Books
      summary: Create a new book
      requestBody:
        description: Details of the new book
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewBook"
      responses:
        "201":
          description: Book created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"
  /api/books/{bookId}:
    get:
      operationId: GetBookById::BookService
      x-route-group: bookGroup::/api/books
      tags:
        - Books
      summary: Retrieve details of a specific book
      parameters:
        - name: bookId
          in: path
          required: true
          description: The ID of the book
          schema:
            type: string
      responses:
        "200":
          description: Details of the book
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"
        "404":
          description: Book not found
    put:
      operationId: UpdateBook::BookService
      x-route-group: bookGroup::/api/books
      tags:
        - Books
      summary: Update details of a specific book
      parameters:
        - name: bookId
          in: path
          required: true
          description: The ID of the book
          schema:
            type: string
      requestBody:
        description: Updated details of the book
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewBook"
      responses:
        "200":
          description: Book updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"
        "404":
          description: Book not found
    delete:
      operationId: DeleteBook::BookService
      x-route-group: bookGroup::/api/books
      tags:
        - Books
      summary: Delete a specific book
      parameters:
        - name: bookId
          in: path
          required: true
          description: The ID of the book
          schema:
            type: string
      responses:
        "204":
          description: Book deleted successfully
        "404":
          description: Book not found
components:
  schemas:
    Book:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        author:
          type: string
        publishedDate:
          type: string
          format: date
        isbn:
          type: string
        availableCopies:
          type: integer
    NewBook:
      type: object
      properties:
        title:
          type: string
        author:
          type: string
        publishedDate:
          type: string
          format: date
        isbn:
          type: string
        availableCopies:
          type: integer
      required:
        - title
        - author
        - publishedDate
        - isbn
        - availableCopies